+++ crates/runtime/src/accelerated_table/refresh_task.rs
@@ -14,7 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-use arrow::array::{Int32Array, Int64Array, RecordBatch, StringArray};
 use arrow::compute::{filter_record_batch, SortOptions};
 use arrow::{
     array::{make_comparator, StructArray, TimestampNanosecondArray},
@@ -22,9 +21,6 @@ use arrow::{
 };
 use async_stream::stream;
 use cache::QueryResultsCacheProvider;
-use data_components::cdc::{ChangeBatch, ChangesStream};
-use data_components::delete::get_deletion_provider;
-use datafusion::logical_expr::lit;
 use futures::{Stream, StreamExt};
 use snafu::{OptionExt, ResultExt};
 use util::fibonacci_backoff::FibonacciBackoffBuilder;
@@ -61,6 +57,9 @@ use datafusion::{
 use datafusion::{execution::context::SessionContext, physical_plan::collect};
 
 use super::refresh::Refresh;
+
+mod changes;
+
 pub struct RefreshTask {
     dataset_name: TableReference,
     federated: Arc<dyn TableProvider>,
@@ -82,67 +81,6 @@ impl RefreshTask {
             refresh,
             accelerator,
         }
-    }
-
-    pub async fn start_changes_stream(
-        &self,
-        mut changes_stream: ChangesStream,
-        cache_provider: Option<Arc<QueryResultsCacheProvider>>,
-        ready_sender: Option<oneshot::Sender<()>>,
-    ) -> super::Result<()> {
-        self.mark_dataset_status(status::ComponentStatus::Refreshing)
-            .await;
-
-        let dataset_name = self.dataset_name.clone();
-
-        let mut ready_sender = ready_sender;
-
-        while let Some(update) = changes_stream.next().await {
-            match update {
-                Ok(change_envelope) => {
-                    match self
-                        .write_change(change_envelope.change_batch.clone())
-                        .await
-                    {
-                        Ok(()) => {
-                            if let Some(ready_sender) = ready_sender.take() {
-                                ready_sender.send(()).ok();
-                            }
-
-                            if let Err(e) = change_envelope.commit() {
-                                tracing::debug!("Failed to commit CDC change envelope: {e}");
-                            }
-
-                            if let Some(cache_provider) = &cache_provider {
-                                if let Err(e) = cache_provider
-                                    .invalidate_for_table(&dataset_name.to_string())
-                                    .await
-                                {
-                                    tracing::error!(
-                                        "Failed to invalidate cached results for dataset {}: {e}",
-                                        &dataset_name.to_string()
-                                    );
-                                }
-                            }
-                        }
-                        Err(e) => {
-                            self.mark_dataset_status(status::ComponentStatus::Error)
-                                .await;
-                            tracing::error!("Error writing change for {dataset_name}: {e}");
-                        }
-                    }
-                }
-                Err(e) => {
-                    tracing::error!("Changes stream error for {dataset_name}: {e}");
-                    self.mark_dataset_status(status::ComponentStatus::Error)
-                        .await;
-                }
-            }
-        }
-
-        tracing::warn!("Changes stream ended for dataset {dataset_name}");
-
-        Ok(())
     }
 
     pub async fn start_streaming_append(
@@ -261,202 +199,6 @@ impl RefreshTask {
         };
 
         self.get_data_update(filters).await
-    }
-
-    async fn write_change(&self, change_batch: ChangeBatch) -> super::Result<()> {
-        let dataset_name = self.dataset_name.clone();
-        let deletion_provider = get_deletion_provider(Arc::clone(&self.accelerator))
-            .context(super::AcceleratedTableDoesntSupportDeleteSnafu)?;
-
-        for row in 0..change_batch.record.num_rows() {
-            let op = change_batch.op(row);
-            match op {
-                "d" => {
-                    let inner_data: RecordBatch = change_batch.data(row);
-                    let primary_keys = change_batch.primary_keys(row);
-                    let delete_where_exprs =
-                        Self::get_delete_where_expr(&inner_data, primary_keys)?;
-
-                    tracing::info!(
-                        "Deleting data for {dataset_name} where {}",
-                        delete_where_exprs
-                            .iter()
-                            .map(ToString::to_string)
-                            .collect::<Vec<_>>()
-                            .join(" AND ")
-                    );
-
-                    let ctx = SessionContext::new();
-                    let session_state = ctx.state();
-
-                    let delete_plan = deletion_provider
-                        .delete_from(&session_state, &delete_where_exprs)
-                        .await
-                        .context(super::FailedToWriteDataSnafu)?;
-
-                    collect(delete_plan, ctx.task_ctx())
-                        .await
-                        .context(super::FailedToWriteDataSnafu)?;
-                }
-                "c" | "u" | "r" => {
-                    let inner_data: RecordBatch = change_batch.data(row);
-                    let primary_keys = change_batch.primary_keys(row);
-                    let ctx = SessionContext::new();
-                    let session_state = ctx.state();
-
-                    tracing::info!(
-                        "Inserting data row for {dataset_name} with {}",
-                        Self::get_primary_key_log_fmt(&inner_data, &primary_keys)?
-                    );
-
-                    let insert_plan = self
-                        .accelerator
-                        .insert_into(
-                            &session_state,
-                            Arc::new(DataUpdateExecutionPlan::new(DataUpdate {
-                                schema: inner_data.schema(),
-                                data: vec![inner_data],
-                                update_type: UpdateType::Append,
-                            })),
-                            false,
-                        )
-                        .await
-                        .context(super::FailedToWriteDataSnafu)?;
-
-                    collect(insert_plan, ctx.task_ctx())
-                        .await
-                        .context(super::FailedToWriteDataSnafu)?;
-                }
-                _ => {
-                    tracing::error!("Unknown operation {op} for {dataset_name}");
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    fn get_primary_key_log_fmt(
-        data: &RecordBatch,
-        primary_keys: &[String],
-    ) -> super::Result<String> {
-        let data_schema = data.schema();
-        primary_keys
-            .iter()
-            .map(|key| {
-                let Some((primary_key_idx, field)) = data_schema.column_with_name(key) else {
-                    return super::ExpectedSchemaToHaveFieldSnafu {
-                        field_name: key.clone(),
-                        schema: Arc::clone(&data_schema),
-                    }
-                    .fail();
-                };
-                let key_col = data.column(primary_key_idx);
-                match field.data_type() {
-                    DataType::Int32 => {
-                        let key_col = key_col.as_any().downcast_ref::<Int32Array>().context(
-                            super::ArrayDataTypeMismatchSnafu {
-                                field_name: key.clone(),
-                                expected_data_type: "Int32".to_string(),
-                                schema: Arc::clone(&data_schema),
-                            },
-                        )?;
-                        Ok(format!("{key}={value}", value = key_col.value(0)))
-                    }
-                    DataType::Int64 => {
-                        let key_col = key_col.as_any().downcast_ref::<Int64Array>().context(
-                            super::ArrayDataTypeMismatchSnafu {
-                                field_name: key.clone(),
-                                expected_data_type: "Int64".to_string(),
-                                schema: Arc::clone(&data_schema),
-                            },
-                        )?;
-                        Ok(format!(
-                            "{key}={value}",
-                            key = key,
-                            value = key_col.value(0)
-                        ))
-                    }
-                    DataType::Utf8 => {
-                        let key_col = key_col.as_any().downcast_ref::<StringArray>().context(
-                            super::ArrayDataTypeMismatchSnafu {
-                                field_name: key.clone(),
-                                expected_data_type: "String".to_string(),
-                                schema: Arc::clone(&data_schema),
-                            },
-                        )?;
-                        Ok(format!(
-                            "{key}={value}",
-                            key = key,
-                            value = key_col.value(0)
-                        ))
-                    }
-                    _ => super::PrimaryKeyTypeNotYetSupportedSnafu {
-                        data_type: field.data_type().to_string(),
-                    }
-                    .fail(),
-                }
-            })
-            .collect::<super::Result<Vec<String>>>()
-            .map(|keys| keys.join(", "))
-    }
-
-    fn get_delete_where_expr(
-        data: &RecordBatch,
-        primary_keys: Vec<String>,
-    ) -> super::Result<Vec<Expr>> {
-        let mut delete_where_exprs: Vec<Expr> = vec![];
-        let data_schema = data.schema();
-        for primary_key in primary_keys {
-            let Some((primary_key_idx, field)) = data_schema.column_with_name(&primary_key) else {
-                return super::ExpectedSchemaToHaveFieldSnafu {
-                    field_name: primary_key.clone(),
-                    schema: Arc::clone(&data_schema),
-                }
-                .fail();
-            };
-            let key_col = data.column(primary_key_idx);
-            let delete_where_expr = match field.data_type() {
-                DataType::Int32 => {
-                    let key_col = key_col.as_any().downcast_ref::<Int32Array>().context(
-                        super::ArrayDataTypeMismatchSnafu {
-                            field_name: primary_key.clone(),
-                            expected_data_type: "Int32".to_string(),
-                            schema: Arc::clone(&data_schema),
-                        },
-                    )?;
-                    col(primary_key).eq(lit(key_col.value(0)))
-                }
-                DataType::Int64 => {
-                    let key_col = key_col.as_any().downcast_ref::<Int64Array>().context(
-                        super::ArrayDataTypeMismatchSnafu {
-                            field_name: primary_key.clone(),
-                            expected_data_type: "Int64".to_string(),
-                            schema: Arc::clone(&data_schema),
-                        },
-                    )?;
-                    col(primary_key).eq(lit(key_col.value(0)))
-                }
-                DataType::Utf8 => {
-                    let key_col = key_col.as_any().downcast_ref::<StringArray>().context(
-                        super::ArrayDataTypeMismatchSnafu {
-                            field_name: primary_key.clone(),
-                            expected_data_type: "String".to_string(),
-                            schema: Arc::clone(&data_schema),
-                        },
-                    )?;
-                    col(primary_key).eq(lit(key_col.value(0)))
-                }
-                _ => {
-                    return super::PrimaryKeyTypeNotYetSupportedSnafu {
-                        data_type: field.data_type().to_string(),
-                    }
-                    .fail()
-                }
-            };
-            delete_where_exprs.push(delete_where_expr);
-        }
-        Ok(delete_where_exprs)
     }
 
     async fn write_data_update(
